+++ sdk/metric/aggregator/exponential/mapping/exponent/exponent.go
@@ -19,6 +19,7 @@ import (
 	"math"
 
 	"go.opentelemetry.io/otel/sdk/metric/aggregator/exponential/mapping"
+	"go.opentelemetry.io/otel/sdk/metric/aggregator/exponential/mapping/internal"
 )
 
 const (
@@ -63,52 +64,61 @@ func NewMapping(scale int32) (mapping.Mapping, error) {
 	return &prebuiltMappings[scale-MinScale], nil
 }
 
+// minNormalLowerBoundaryIndex is the largest index such that
+// base**index is <= MinValue.  A histogram bucket with this index
+// covers the range (base**index, base**(index+1)], including
+// MinValue.
+func (e *exponentMapping) minNormalLowerBoundaryIndex() int32 {
+	idx := int32(internal.MinNormalExponent) >> e.shift
+	if e.shift < 2 {
+		// For scales -1 and 0 the minimum value 2**-1022
+		// is a power-of-two multiple, meaning it belongs
+		// to the index one less.
+		idx--
+	}
+	return idx
+}
+
+// maxNormalLowerBoundaryIndex is the index such that base**index
+// equals the largest representable boundary.  A histogram bucket with this
+// index covers the range (0x1p+1024/base, 0x1p+1024], which includes
+// MaxValue; note that this bucket is incomplete, since the upper
+// boundary cannot be represented.  One greater than this index
+// corresponds with the bucket containing values > 0x1p1024.
+func (e *exponentMapping) maxNormalLowerBoundaryIndex() int32 {
+	return int32(internal.MaxNormalExponent) >> e.shift
+}
+
 // MapToIndex implements mapping.Mapping.
 func (e *exponentMapping) MapToIndex(value float64) int32 {
 	// Note: we can assume not a 0, Inf, or NaN; positive sign bit.
+	if value < internal.MinValue {
+		return e.minNormalLowerBoundaryIndex()
+	}
+
+	// Extract the raw exponent.
+	rawExp := internal.GetNormalBase2(value)
+
+	// In case the value is an exact power of two, compute a
+	// correction of -1:
+	correction := int32((internal.GetSignificand(value) - 1) >> internal.SignificandWidth)
 
 	// Note: bit-shifting does the right thing for negative
 	// exponents, e.g., -1 >> 1 == -1.
-	return getBase2(value) >> e.shift
-}
-
-func (e *exponentMapping) minIndex() int32 {
-	return int32(MinNormalExponent) >> e.shift
-}
-
-func (e *exponentMapping) maxIndex() int32 {
-	return int32(MaxNormalExponent) >> e.shift
+	return (rawExp + correction) >> e.shift
 }
 
 // LowerBoundary implements mapping.Mapping.
 func (e *exponentMapping) LowerBoundary(index int32) (float64, error) {
-	if min := e.minIndex(); index < min {
+	if min := e.minNormalLowerBoundaryIndex(); index < min {
 		return 0, mapping.ErrUnderflow
 	}
 
-	if max := e.maxIndex(); index > max {
+	if max := e.maxNormalLowerBoundaryIndex(); index > max {
 		return 0, mapping.ErrOverflow
 	}
 
-	unbiased := int64(index << e.shift)
-
-	// Note: although the mapping function rounds subnormal values
-	// up to the smallest normal value, there are still buckets
-	// that may be filled that start at subnormal values.  The
-	// following code handles this correctly.  It's equivalent to and
-	// faster than math.Ldexp(1, int(unbiased)).
-	if unbiased < int64(MinNormalExponent) {
-		subnormal := uint64(1 << SignificandWidth)
-		for unbiased < int64(MinNormalExponent) {
-			unbiased++
-			subnormal >>= 1
-		}
-		return math.Float64frombits(subnormal), nil
-	}
-	exponent := unbiased + ExponentBias
-
-	bits := uint64(exponent << SignificandWidth)
-	return math.Float64frombits(bits), nil
+	return math.Ldexp(1, int(index<<e.shift)), nil
 }
 
 // Scale implements mapping.Mapping.